refactor: move `clear_logs` to `logging` module
